Slim Bat logging and recording code
